+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.3
 
 - add car type into name
+- added port setting
 
 # 1.0.2